Extract Basiclayout to a separate file
